+++ maintenance-core/src/main/java/eu/kennytv/maintenance/core/hook/ServerListPlusHook.java
@@ -25,8 +25,9 @@ public final class ServerListPlusHook {
     private final ServerListPlusCore serverListPlusCore;
 
     public ServerListPlusHook(final Object serverListPlus) {
-        if (!(serverListPlus instanceof ServerListPlusPlugin))
+        if (!(serverListPlus instanceof ServerListPlusPlugin)) {
             throw new IllegalArgumentException("Couldn't parse ServerListPlus instance!");
+        }
         this.serverListPlusCore = ((ServerListPlusPlugin) serverListPlus).getCore();
     }