+++ maintenance-velocity/src/main/java/eu/kennytv/maintenance/velocity/util/VelocityServer.java
@@ -21,8 +21,8 @@ package eu.kennytv.maintenance.velocity.util;
 import com.velocitypowered.api.proxy.Player;
 import com.velocitypowered.api.proxy.server.RegisteredServer;
 import eu.kennytv.maintenance.api.proxy.Server;
-import net.kyori.text.TextComponent;
-import net.kyori.text.serializer.legacy.LegacyComponentSerializer;
+import net.kyori.adventure.text.TextComponent;
+import net.kyori.adventure.text.serializer.legacy.LegacyComponentSerializer;
 
 public final class VelocityServer implements Server {
     private final RegisteredServer server;
@@ -43,7 +43,7 @@ public final class VelocityServer implements Server {
 
     @Override
     public void broadcast(final String message) {
-        final TextComponent s = LegacyComponentSerializer.INSTANCE.deserialize(message);
+        final TextComponent s = LegacyComponentSerializer.legacySection().deserialize(message);
         for (final Player p : server.getPlayersConnected()) {
             p.sendMessage(s);
         }